Create find-duplicate-in-array.py
Given a read only array of n + 1 integers between 1 and n, find one number that repeats in linear time using less than O(n) space and traversing the stream sequentially O(1) times.

Solution:
While traversing the array once, use the absolute value of every element as the index and make the value at this index negative to mark it as visited. If something is already negative it means that the element has repeated.